add back helper fns to lib
